split tests into files and change flake8/pytest setup
* extends tox and travis config to more python versions
* lines up closer with pep8
