+++ setup.py
@@ -43,5 +43,6 @@ def main():
         py_modules=['pluggy'],
     )
 
+
 if __name__ == '__main__':
     main()
