+++ tox.ini
@@ -1,23 +1,21 @@
 [tox]
-envlist=check,{py26,py27,py34,py35,pypy}-{pytest27,pytest}
+envlist=check,py{26,27,34,35,py}-pytest{28,29,30}
 
 [testenv]
-commands= py.test {posargs:testing/test_pluggy.py}
+commands= py.test {posargs:testing/}
 deps=
-    pytest27: pytest>=2.7.0,<2.8.0
-    pytest: pytest>=2.8.0
+  pytest28: pytest~=2.8.0
+  pytest29: pytest~=2.9.0
+  pytest30: pytest~=3.0.0
 
 
 [testenv:check]
-usedevelop=True
-deps = pytest>=2.7.0,<2.8.0
-       pytest-pep8
-       pytest-flakes
-       restructuredtext_lint 
-commands =
-    py.test --pep8
-    py.test --flakes -m flakes pluggy.py testing/test_pluggy.py
-    rst-lint CHANGELOG.rst README.rst
+deps =
+  flake8
+  restructuredtext_lint 
+commands = 
+  flake8 pluggy.py setup.py testing
+  rst-lint CHANGELOG.rst README.rst
 
 
 [testenv:docs]
@@ -33,5 +31,7 @@ commands =
 minversion=2.0
 #--pyargs --doctest-modules --ignore=.tox
 addopts= -rxsX 
-pep8ignore = E501 E128 E127
 norecursedirs = .tox ja .hg .env*
+
+[flake8]
+max-line-length = 99
